Move kanban models to their own files
